Split openapi2 package into files a la openapi3/
